Implemented gyro and touch controls.
For debugging purposes, I used the following script:
http://wiki.unity3d.com/index.php/DebugConsole
For the gyro, there is a deadzone and gradual increase in side to side
speed based on the angle. The angle is clamped at 30 degrees so the player
doesn't have to tilt the device too much to get maximum sideways turn
power.