Fixed bouncing problem and started work on shadows.
To fix the bouncing, the collider was removed from the slope chunks and
instead a super long, invisible slope was created with the collider. This
is still not that inefficient because the rendering occurs in chunks.

For the shadows, the included shaders DO work, but currently aren't
playing nice in Unity.