Accelerometer, improved logic and tree hit penalty.
The player now loses 50 points for hitting a tree to stop them cheating by
using trees to slow down. Player movement is now controlled by the
accelerometer, not the gyroscope. Furthermore, movement code has been
fixed. Previously, the forward speed was attached to the sideways speed.
The majority of the movement code is also now in FixedUpdate so the frame
rate doesn't affect the physics.